Support for double numbers

+++ exponential/Model/AccountBalance.swift
@@ -11,28 +11,28 @@ import Foundation
 struct AccountBalance :  Codable {
     
     init() {
-        current = Int.random(in: 1..<20000)
-        limit = Int.random(in: 1..<20000) + current!
-        available = Int.random(in: 1..<15000)
+        current = Double(Int.random(in: 1..<20000))
+        limit = Double(Int.random(in: 1..<20000)) + current!
+        available = Double(Int.random(in: 1..<15000))
     }
     
     init(_ balances: NSDictionary) {
-        available = balances["available"] as? Int
-        limit = balances["limit"] as? Int
-        current = balances["current"] as? Int
+        available = balances["available"] as? Double
+        limit = balances["limit"] as? Double
+        current = balances["current"] as? Double
         iso_currency_code = balances["iso_currency_code"] as? String
         unofficial_currency_code = balances["unofficial_currency_code"] as? String
     }
     
-    func getCurrent() -> Int {OptUtils.unwrap(current)}
-    func getAvailable() -> Int {OptUtils.unwrap(available)}
-    func getLimit() -> Int {OptUtils.unwrap(limit)}
+    func getCurrent() -> Double {OptUtils.unwrap(current)}
+    func getAvailable() -> Double {OptUtils.unwrap(available)}
+    func getLimit() -> Double {OptUtils.unwrap(limit)}
     func getCurrencyCode() -> String {OptUtils.unwrap(iso_currency_code)}
     func getUnofficialCurrencyCode() -> String {OptUtils.unwrap(unofficial_currency_code)}
     
     private var iso_currency_code: String?
     private var unofficial_currency_code: String?
-    private var current : Int?
-    private var limit : Int?
-    private var available : Int?
+    private var current : Double?
+    private var limit : Double?
+    private var available : Double?
 }
